+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Bump dependency versions
 - Add map(), filter(), any(), all() to ArrayCollection and ResultCollection in ORM
+- Remove deprecated Attributes from GraphQL extension
 
 ## 0.2.15